removing all duplications in misc.hpp
