+++ cpp/oneapi/dal/algo/covariance/backend/gpu/finalize_compute_kernel_dense_dpc.cpp
@@ -15,6 +15,7 @@
 *******************************************************************************/
 
 #include "oneapi/dal/algo/covariance/backend/gpu/finalize_compute_kernel.hpp"
+#include "oneapi/dal/algo/covariance/backend/gpu/misc.hpp"
 
 #include "oneapi/dal/backend/primitives/lapack.hpp"
 #include "oneapi/dal/backend/primitives/reduction.hpp"
@@ -33,7 +34,7 @@ using bk::context_gpu;
 using task_t = task::compute;
 using input_t = partial_compute_result<task_t>;
 using result_t = compute_result<task_t>;
-using descriptor_t = detail::descriptor_base<task::compute>;
+using descriptor_t = detail::descriptor_base<task_t>;
 
 template <typename Float, typename Task>
 static compute_result<Task> finalize_compute(const context_gpu& ctx,
@@ -60,55 +61,33 @@ static compute_result<Task> finalize_compute(const context_gpu& ctx,
     const auto xtx =
         pr::table2ndarray<Float>(q, input.get_partial_crossproduct(), sycl::usm::alloc::device);
 
-    sycl::event matrix_event;
     if (desc.get_result_options().test(result_options::cov_matrix)) {
-        auto cov = pr::ndarray<Float, 2>::empty(q, { column_count, column_count }, alloc::device);
-        sycl::event copy_event;
-        {
-            ONEDAL_PROFILER_TASK(copy_cov, q);
-            copy_event = copy(q, cov, xtx, {});
-        }
-
-        {
-            ONEDAL_PROFILER_TASK(compute_covariance_matrix, q);
-            matrix_event = pr::covariance(q, rows_count_global, sums, cov, bias, { copy_event });
-        }
+        auto [cov, cov_event] = compute_covariance(q, rows_count_global, xtx, sums, bias);
         result.set_cov_matrix(
-            (homogen_table::wrap(cov.flatten(q, { matrix_event }), column_count, column_count)));
+            (homogen_table::wrap(cov.flatten(q, { cov_event }), column_count, column_count)));
     }
     if (desc.get_result_options().test(result_options::cor_matrix)) {
-        auto corr = pr::ndarray<Float, 2>::empty(q, { column_count, column_count }, alloc::device);
-
-        sycl::event copy_corr_event;
-        {
-            ONEDAL_PROFILER_TASK(copy_corr, q);
-            copy_corr_event = copy(q, corr, xtx, {});
-        }
-        {
-            ONEDAL_PROFILER_TASK(compute_covariance_matrix, q);
-            matrix_event = pr::correlation(q, rows_count_global, sums, corr, { copy_corr_event });
-        }
+        auto [corr, corr_event] = compute_correlation(q, rows_count_global, xtx, sums);
         result.set_cor_matrix(
-            (homogen_table::wrap(corr.flatten(q, { matrix_event }), column_count, column_count)));
+            (homogen_table::wrap(corr.flatten(q, { corr_event }), column_count, column_count)));
     }
     if (desc.get_result_options().test(result_options::means)) {
-        auto means = pr::ndarray<Float, 1>::empty(q, { column_count }, alloc::device);
-        auto means_event = pr::means(q, rows_count_global, sums, means, { matrix_event });
+        auto [means, means_event] = compute_means(q, sums, rows_count_global);
         result.set_means(homogen_table::wrap(means.flatten(q), 1, column_count));
     }
     return result;
 }
 
 template <typename Float>
-struct finalize_compute_kernel_gpu<Float, method::dense, task::compute> {
+struct finalize_compute_kernel_gpu<Float, method::dense, task_t> {
     result_t operator()(const context_gpu& ctx,
                         const descriptor_t& desc,
                         const input_t& input) const {
-        return finalize_compute<Float, task::compute>(ctx, desc, input);
+        return finalize_compute<Float, task_t>(ctx, desc, input);
     }
 };
 
-template struct finalize_compute_kernel_gpu<float, method::dense, task::compute>;
-template struct finalize_compute_kernel_gpu<double, method::dense, task::compute>;
+template struct finalize_compute_kernel_gpu<float, method::dense, task_t>;
+template struct finalize_compute_kernel_gpu<double, method::dense, task_t>;
 
 } // namespace oneapi::dal::covariance::backend
